requirements.txt
Flask>=2.2.5
Flask-SocketIO>=5.2.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-Mail>=0.9.1
Werkzeug>=2.2.2
python-dotenv>=1.0.0
email-validator>=2.1.0.post1
